Added solution of crash at a public endpoint after an authenticated
It crashes at request of a public endpoint after request of an authenticated endpoint.

Because of using the same curl during DoGETrequest and DoPOSTrequest, but the last injects trash (for DoGETrequest) into curl.

Here is a brilliant solution of this difficult for solving problem, so solution must be included at example.cpp because it's an example.